Combine import wizards into one
