+++ kolibri/plugins/device_management/assets/src/views/manage-content-page/select-transfer-source-modal/select-drive-modal.vue
